@@ -2,11 +2,13 @@
 
 import os
 import json
+from pathlib import Path
 
 import docutils
 from docutils.parsers.rst import Directive, directives
-from docutils.nodes import math_block, image
+from docutils.nodes import math_block, image, literal
 from sphinx.util import parselinenos
+from sphinx.util.docutils import ReferenceRole
 from sphinx.addnodes import download_reference
 from sphinx.transforms import SphinxTransform
 from sphinx.environment.collectors.asset import ImageCollector
@@ -100,7 +102,7 @@ class JupyterCell(Directive):
                     location=location,
                 )
             try:
-                with open(filename) as f:
+                with Path(filename).open() as f:
                     content = [line.rstrip() for line in f.readlines()]
             except (IOError, OSError):
                 raise IOError("File {} not found or reading it failed".format(filename))
@@ -227,7 +229,7 @@ class JupyterWidgetStateNode(docutils.nodes.Element):
         )
 
 
-def cell_output_to_nodes(outputs, data_priority, write_stderr, dir,
+def cell_output_to_nodes(outputs, data_priority, write_stderr, out_dir,
                          thebe_config, inline=False):
     """Convert a jupyter cell with outputs and filenames to doctree nodes.
 
@@ -238,7 +240,7 @@ def cell_output_to_nodes(outputs, data_priority, write_stderr, dir,
         Which media types to prioritize.
     write_stderr : bool
         If True include stderr in cell output
-    dir : string
+    out_dir : string
         Sphinx "absolute path" to the output folder, so it is a relative path
         to the source folder prefixed with ``/``.
     thebe_config: dict
@@ -317,19 +319,17 @@ def cell_output_to_nodes(outputs, data_priority, write_stderr, dir,
                 continue
             data = output["data"][mime_type]
             if mime_type.startswith("image"):
+                file_path = Path(output.metadata["filenames"][mime_type])
+                out_dir = Path(out_dir)
                 # Sphinx treats absolute paths as being rooted at the source
                 # directory, so make a relative path, which Sphinx treats
                 # as being relative to the current working directory.
-                filename = os.path.basename(output.metadata["filenames"][mime_type])
+                filename = file_path.name
 
-                # checks if file dir path is inside a subdir of dir
-                filedir = os.path.dirname(output.metadata["filenames"][mime_type])
-                subpaths = filedir.split(dir)
-                if subpaths and len(subpaths) > 1:
-                    subpath = subpaths[1]
-                    dir += subpath
+                if out_dir in file_path.parents:
+                    out_dir = file_path.parent
 
-                uri = os.path.join(dir, filename)
+                uri = (out_dir / filename).as_posix()
                 to_add.append(docutils.nodes.image(uri=uri))
             elif mime_type == "text/html":
                 to_add.append(
@@ -407,16 +407,19 @@ def attach_outputs(output_nodes, node, thebe_config, cm_language):
         node.children = node.children[::-1]
 
 
-def jupyter_download_role(name, rawtext, text, lineno, inliner):
-    _, filetype = name.split(":")
-    assert filetype in ("notebook", "script")
-    ext = ".ipynb" if filetype == "notebook" else ".py"
-    output_dir = sphinx_abs_dir(inliner.document.settings.env)
-    download_file = text + ext
-    node = download_reference(
-        download_file, download_file, reftarget=os.path.join(output_dir, download_file)
-    )
-    return [node], []
+class JupyterDownloadRole(ReferenceRole):
+    def run(self):
+        _, filetype = self.name.split(":")
+
+        assert filetype in ("notebook", "nb", "script")
+        ext = ".ipynb" if filetype in ("notebook", "nb") else ".py"
+        download_file = self.target + ext
+        reftarget = sphinx_abs_dir(self.env, download_file)
+        node = download_reference(self.rawtext, reftarget=reftarget)
+        self.set_source_info(node)
+        title = self.title if self.has_explicit_title else download_file
+        node += literal(self.rawtext, title, classes=["xref", "download"])
+        return [node], []
 
 
 def get_widgets(notebook):
